Fixed Avavlon reads
The CoCoTb Avalon master driver and corsair do not agree with one and
other.

The driver removes the addres one a read pulse has been asserted (1
clock cycle) while the hdl of corsair expects it to remain stable for at
least one more cycle.
This has been fixed by clocking in the address at a read pulse and use
that for the read process.